+++ packages/react-hooks/src/index.ts
@@ -34,3 +34,4 @@ export * from './useSetAttributesCallback';
 export * from './useSpectrumDisableSpellcheckRef';
 export * from './useWindowedListData';
 export * from './useResizeObserver';
+export * from './useMergeRef';
